refactor(tests): move NVOS polling tests to tested code mod
