+++ crates/api-core/src/tests/rack_state_controller/handler.rs
@@ -22,7 +22,6 @@ use carbide_rack_controller::config::ScaleUpFabricManagerApiVersion;
 use carbide_rack_controller::context::RackStateHandlerContextObjects;
 use carbide_rack_controller::firmware_object::FirmwareObjectFetcher;
 use carbide_rack_controller::handler::RackStateHandler;
-use carbide_rack_controller::maintenance::apply_nvos_job_status_response;
 use carbide_rack_controller::metrics::RackMetrics;
 use carbide_secrets::credentials::{
     BmcCredentialType, CredentialKey, CredentialReader, Credentials,
@@ -40,9 +39,9 @@ use model::expected_rack::ExpectedRack;
 use model::rack::{
     ConfigureNmxClusterCertificateState, ConfigureNmxClusterState, FirmwareUpgradeDeviceStatus,
     FirmwareUpgradeJob, FirmwareUpgradeState, MaintenanceActivity, MaintenanceScope,
-    NvosUpdateState, NvosUpdateSwitchStatus, Rack, RackConfig, RackFirmwareUpgradeState,
-    RackFirmwareUpgradeStatus, RackMaintenanceState, RackPowerState, RackState,
-    RackValidationState, SwitchNvosUpdateState, SwitchNvosUpdateStatus,
+    NvosUpdateState, Rack, RackConfig, RackFirmwareUpgradeState, RackFirmwareUpgradeStatus,
+    RackMaintenanceState, RackPowerState, RackState, RackValidationState, SwitchNvosUpdateState,
+    SwitchNvosUpdateStatus,
 };
 use model::rack_type::{
     RackCapabilitiesSet, RackCapabilityCompute, RackCapabilityPowerShelf, RackCapabilitySwitch,
@@ -856,93 +855,6 @@ async fn test_expected_no_definition_stays_parked(
     );
 
     Ok(())
-}
-
-#[test]
-fn test_nvos_polling_updates_node_id_and_maps_running_to_in_progress() {
-    let mut switch = NvosUpdateSwitchStatus {
-        node_id: "old-node-id".into(),
-        mac: "00:11:22:33:44:55".into(),
-        bmc_ip: "10.0.0.10".into(),
-        nvos_ip: "192.168.10.10".into(),
-        status: "pending".into(),
-        job_id: Some("job-1".into()),
-        error_message: Some("stale error".into()),
-    };
-
-    apply_nvos_job_status_response(
-        &mut switch,
-        "job-1",
-        Ok(rms::GetSwitchSystemImageJobStatusResponse {
-            status: rms::ReturnCode::Success as i32,
-            state: "RUNNING".into(),
-            node_id: "new-node-id".into(),
-            ..Default::default()
-        }),
-    );
-
-    assert_eq!(switch.node_id, "new-node-id");
-    assert_eq!(switch.status, "in_progress");
-    assert_eq!(switch.error_message, None);
-}
-
-#[test]
-fn test_nvos_polling_maps_failed_state_and_uses_error_message() {
-    let mut switch = NvosUpdateSwitchStatus {
-        node_id: "node-id".into(),
-        mac: "00:11:22:33:44:55".into(),
-        bmc_ip: "10.0.0.10".into(),
-        nvos_ip: "192.168.10.10".into(),
-        status: "in_progress".into(),
-        job_id: Some("job-2".into()),
-        error_message: None,
-    };
-
-    apply_nvos_job_status_response(
-        &mut switch,
-        "job-2",
-        Ok(rms::GetSwitchSystemImageJobStatusResponse {
-            status: rms::ReturnCode::Success as i32,
-            state: "failed".into(),
-            error_message: "image install failed".into(),
-            ..Default::default()
-        }),
-    );
-
-    assert_eq!(switch.status, "failed");
-    assert_eq!(
-        switch.error_message.as_deref(),
-        Some("image install failed")
-    );
-}
-
-#[test]
-fn test_nvos_polling_unknown_state_preserves_status_and_sets_error() {
-    let mut switch = NvosUpdateSwitchStatus {
-        node_id: "node-id".into(),
-        mac: "00:11:22:33:44:55".into(),
-        bmc_ip: "10.0.0.10".into(),
-        nvos_ip: "192.168.10.10".into(),
-        status: "pending".into(),
-        job_id: Some("job-3".into()),
-        error_message: None,
-    };
-
-    apply_nvos_job_status_response(
-        &mut switch,
-        "job-3",
-        Ok(rms::GetSwitchSystemImageJobStatusResponse {
-            status: rms::ReturnCode::Success as i32,
-            state: "mystery".into(),
-            ..Default::default()
-        }),
-    );
-
-    assert_eq!(switch.status, "pending");
-    assert_eq!(
-        switch.error_message.as_deref(),
-        Some("Unknown RMS switch image job state mystery")
-    );
 }
 
 /// test_expected_incomplete_device_counts_stays verifies that a rack with a
